Move sketch to appropriately named subfolder
The Arduino IDE requires that a sketch be located in a folder of the same name. Although the name of the repository does match the sketch name, when GitHub's popular Clone or download > Download ZIP feature is used to download the contents of a repository the branch/tag/commit name is appended to the folder name, causing a mismatch.

When opening a file that does not meet this sketch/folder name matching requirement the Arduino IDE presents a dialog:

The file "nofrendo_arcata.ino" needs to be inside a sketch folder named "nofrendo_arcata".
Create this folder, move the file, and continue?

After clicking "OK" the Arduino IDE currently moves only the file nofrendo_arcata.ino to the new folder, leaving behind the other source files. This causes compilation of the sketch to fail:

nofrendo_arcada:2:12: error: emuapi.h: No such file or directory

   #include "emuapi.h"