Move files to subfolder
The src subfolder of the sketch is compiled recursively. The Arduino IDE only opens files from the root folder of the sketch. This allows the source files to be moved out of the root folder of the sketch so that the user is only presented with nofrendo_arcada.ino when they open the sketch. It also makes it easier to navigate through the files of the repository. The mapnnn.c files were moved to the src/mappers folder.